fixes + tests (+ move all run tests to a dir)
